CellFreeModelGenerationKit.jl

A package for building cell-free models in Julia language. The purpose of this package is to generate model code for performing constraint-based modeling like flux balance anlaysis (FBA).

Getting Started: Installation and Requirements

CellFreeModelGenerationKit.jl is open source, available under a MIT software license. You can download this repository as a zip file, clone or pull it by using the command (from the command-line):

$ git clone https://github.com/varnerlab/CellFreeModelGenerationKit.git

CellFreeModelGenerationKit.jl is organized as a Julia package which can be installed in the package mode of Julia. Open the Julia REPL and enter the package mode using the ] key. Then, at the prompt enter:

(@v1.5) pkg> add https://github.com/varnerlab/CellFreeModelGenerationKit.git

This will install the CellFreeModelGenerationKit.jl package and the other required packages. CellFreeModelGenerationKit.jl requires Julia 1.5.x and above. There are several other packages that are required to run the model. However, these should be installed automatically the first time you run your code. The linear programming problem is solved using the GLPK solver, which is freely available for a variety of platforms. The following dependencies are installed: DataFrames, CSV, Dates, Logging, WordTokenizers, DelimitedFiles and SQLite.

Methods

CellFreeModelGenerationKit.build_control_program_component — Method

Function: build_control_program_component(intermediate_dictionary::Dict{String,Any})::VLResult

Generate Control.jl. Control.jl includes methods to calculate transcription and translation control functions at time t.

Input arguments:
`intermediate_dictionary::Dict{String,Any}` - data dictionary containing the master reaction table and molecular species participating in the reactions.

Output arguments:
`VLResult::VLResult` - concrete data type holding control values for each rate.

CellFreeModelGenerationKit.build_data_dictionary_program_component — Method

build_data_dictionary_program_component(intermediate_dictionary::Dict{String,Any})::VLResult

Generate Data.jl. Data.jl includes a method to build a default data dictionary. This dictionary will contain information about the model system along with default bounds useful for flux balance analysis.

Input arguments:
`intermediate_dictionary::Dict{String,Any}` - data dictionary containing the master reaction table and molecular species participating in the reactions.

Output arguments:
`VLResult::VLResult` - concrete data type holding the data dictionary.

CellFreeModelGenerationKit.build_julia_model_object — Method

build_julia_model_object(path_to_model_file::String, path_to_output_dir::String; 
    defaults_file_name::String="Defaults.toml", model_type::Symbol=:static)::VLResult

Build a static or dynamic data object in Julia language. If a TOML file with default parameter values is not provided by the user, a `Defaults.toml` file is generated automatically. Additionally, the user can edit the generated `Defaults.toml` file with their own values. After generating code, if a directory already exists at the user specified location, it can be deleted or backed-up before new code is written based on user input.

Input arguments:
`path_to_model_file::String` - user-specified path to read the Network.vff file from as well as the Defaults.toml file (if it already exists)
`path_to_output_dir::String` - path to where generated code will be written
`defaults_file_name::String` - name of the TOML file with default parameter values (optional).
`model_type::Symbol` - type of model to generate (static/ dynamic) (optional).

Output arguments:
`model_object::VLResult{VLJuliaModelObject}` - abstract Julia data object holding information for generating model code.

CellFreeModelGenerationKit.build_kinetics_program_component — Method

build_kinetics_dictionary_program_component(intermediate_dictionary::Dict{String,Any})::VLResult

Generate Kinetics.jl. Kinetics.jl includes methods to calculate transcription and translation rates based on biophysical model parameters.

Input arguments:
`intermediate_dictionary::Dict{String,Any}` - data dictionary containing the master reaction table and molecular species participating in the reactions.

Output arguments:
`VLResult::VLResult` - concrete data type holding all rates.

CellFreeModelGenerationKit.generate — Method

generate(julia_model_object::VLJuliaModelObject; 
intermediate_representation_dictionary::Union{Nothing,Dict{String,Any}} = nothing, 
logger::Union{Nothing,SimpleLogger} = nothing)

Generate model code from a Julia data object and write generated code to the user-defined output directory.

Input arguments:
`julia_model_object::VLJuliaModelObject` - abstract data object for generating model code in Julia language.
`intermediate_representation_dictionary::Union{Nothing,Dict{String,Any}}` - an intermediate data dictionary for building model code (optional).
`logger::Union{Nothing,SimpleLogger}` - logger to log information during the code generation process (optional).

Output arguments:
None

CellFreeModelGenerationKit.generate_default_project_file — Method

generate_default_project(path_to_project_dir::String)::VLResult

Generates a default project structure which contains an empty model file and Defaults.toml file.

Input arguments:
`path_to_project_file::String` - path to where you want model code to be generated

Output arguments:
None

CellFreeModelGenerationKit.generate_stoichiometric_matrix — Method

generate_stoichiometric_matrix(intermediate_dictionary::Dict{String,Any})::VLResult

Generate a stoichiometric matrix based on the biochemical model reaction network.

Input arguments:
`intermediate_dictionary::Dict{String,Any}` - data dictionary containing the master reaction table and molecular species participating in the reactions.

Output arguments:
`VLResult::VLResult` - concrete data type holding the generated stoichiometric matrix.

CellFreeModelGenerationKit.parse_vff_bio_types_section — Method

parse_vff_bio_types_section(buffer::Array{String,1})::VLResult

CellFreeModelGenerationKit.parse_vff_grn_section — Method

parse_grn_section(buffer::Array{String,1})::VLResult

CellFreeModelGenerationKit.parse_vff_metabolic_section — Method

parse_vff_metabolic_section(buffer::Array{String,1})::VLResult

CellFreeModelGenerationKit.parse_vff_model_document — Method

parse_vff_model_document(model::VLAbstractModelObject)::VLResult

CellFreeModelGenerationKit.parse_vff_sequence_section — Method

parse_vff_sequence_section(buffer::Array{String,1})::VLResult

CellFreeModelGenerationKit.parse_vff_species_bounds_section — Method

parse_vff_species_bounds_section(buffer::Array{String,1}, metabolic_results_tuple::NamedTuple)::VLResult

CellFreeModelGenerationKit.read_model_document — Method

read_model_document(path_to_file::String; 
        strip_comments::Bool = true)::Array{String,1}

Read each line of the Network.vff input file and extract contents of the file. Comments in the file (beginning by `//` symbol) are excluded by default, which can be overridden by the user.

Input arguments:
`path_to_file::String` - user-specified path where the Network.vff input file can be found and parsed
`strip_comments::Bool` - if true, comment lines (beginning by ‘//’ symbol) are excluded by the parser (optional).

Output arguments:
`buffer::Array{String,1}` - single array holding information extracted from the Network.vff file.

CellFreeModelGenerationKit.reorder_molecular_symbol_array — Method

reorder_molecular_symbol_array(symbol_array::Array{String,1}; 
      callback::Union{Function,Nothing} = nothing)::Array{String,1}

Reorder species symbol names alphabetically (default implementation) unless a callback function is passed, in which case sorting is done based on the user-defined routine.

Input arguments:
`symbol_array::Array{String,1}` - array holding symbol names for all molecular species in the network.
`callback::Union{Function,Nothing}` - user-defined function to reorder species symbol names (optional).

Output arguments:
`symbol_array::Array{String,1}` - sorted molecular species symbol array.

CellFreeModelGenerationKit.reorder_reaction_symbol_array — Method

reorder_reaction_symbol_array(symbol_array::Array{String,1}; 
      callback::Union{Function,Nothing} = nothing)::Array{String,1}

Reorder reaction symbol names alphabetically (default implementation) unless a callback function is passed, in which case sorting is done based on the user-defined routine.

Input arguments:
`symbol_array::Array{String,1}` - array holding symbol names for all reactions in the network.
`callback::Union{Function,Nothing}` - user-defined function to reorder reaction names (optional).

Output arguments:
`symbol_array::Array{String,1}` - sorted reaction symbol array.