Quick start guide

Obtain metabolism file

Create your own metabolism file that describes the system in consideration with the corresponding TXTL sequence, metabolism and the GRN section. This file should be structured in a VFF format as mentioned here. Note the path to this metabolism file.

Generate model code

To generate cell free model code, first load the CellFreeModelGenerationKit.jl package, then build a Julia model object from the REPL using the command build_julia_model_object.

CellFreeModelGenerationKit.build_julia_model_object — Function

build_julia_model_object(path_to_model_file::String, path_to_output_dir::String;
    defaults_file_name::String="Defaults.toml", model_type::Symbol=:static)::VLResult

Build a static or dynamic data object in Julia language. If a TOML file with default parameter values is not provided by the user, a Defaults.toml file is generated automatically. Additionally, the user can edit the generated Defaults.toml file with their own values. After generating code, if a directory already exists at the user specified location, it can be deleted or backed-up before new code is written based on user input.

Input arguments: path_to_model_file::String - user-specified path to read the Network.vff file from as well as the Defaults.toml file (if it already exists) path_to_output_dir::String - path to where generated code will be written defaults_file_name::String - name of the TOML file with default parameter values (optional). model_type::Symbol - type of model to generate (static/ dynamic) (optional).

Output arguments: model_object::VLResult{VLJuliaModelObject} - abstract Julia data object holding information for generating model code.

Note that the path where files will be written should be different from where model file is.

julia> using CellFreeModelGenerationKit
julia> build_result = build_julia_model_object(path_to_test_metabolism_file, path_to_output_dir)
julia> julia_model_object = build_result.value

The build_julia_model_object reads the model file and a defaults TOML file (if it exists) at the user specified path. If a defaults file is not provided by the user, a Defaults.toml file is generated. Additionally, the user can modify the contents of Defaults.toml.

For writing generated code, if a directory already exists at the user specified location, it can be backed-up before new code is written (based on prompted user input). After building a Julia model object, issue the command generate from the REPL.

CellFreeModelGenerationKit.generate — Function

generate(julia_model_object::VLJuliaModelObject; 
intermediate_representation_dictionary::Union{Nothing,Dict{String,Any}} = nothing, 
logger::Union{Nothing,SimpleLogger} = nothing)

Generate model code from a Julia data object and write generated code to the user-defined output directory.

Input arguments: julia_model_object::VLJuliaModelObject - abstract data object for generating model code in Julia language. intermediate_representation_dictionary::Union{Nothing,Dict{String,Any}} - an intermediate data dictionary for building model code (optional). logger::Union{Nothing,SimpleLogger} - logger to log information during the code generation process (optional).

Output arguments: None

julia> generate(julia_model_object)