CellFreeModelGenerationKit.parse_vff_sequence_section — Function

parse_vff_sequence_section(buffer::Array{String,1})::VLResult

Parse contents of the sequence section in the model file and extract amino acid and nucleotide sequences for protein and DNA, respectively. The sequence section is written between the tags #TXTL-SEQUENCE::START and #TXTL-SEQUENCE::STOP.

Input arguments: buffer::Array{String,1} - 1-D array holding information extracted from the model file.

Output arguments: VLResult::VLResult - concrete type holding protein and DNA sequence information.

CellFreeModelGenerationKit.parse_vff_metabolic_section — Function

parse_vff_metabolic_section(buffer::Array{String,1})::VLResult

Parse contents of the metabolic section in the model file and extract all biochemical reactions and molecular species involved in the metabolic network. The metabolic section is written between the tags #METABOLISM::START and #METABOLISM::STOP.

Input arguments: buffer::Array{String,1} - single array holding information extracted from the model file.

Output arguments: VLResult::VLResult - concrete type holding information on reactions and species involved in metabolism.

CellFreeModelGenerationKit.parse_vff_species_bounds_section — Function

parse_vff_species_bounds_section(buffer::Array{String,1}, metabolic_results_tuple::NamedTuple)::VLResult

Parse contents of the model file and extract information on species bounds. The species bounds section is written between the tags #SPECIESBOUNDS::START and #SPECIESBOUNDS::STOP.

Input arguments: buffer::Array{String,1} - 1-D array holding information extracted from the model file.

Output arguments: VLResult::VLResult - concrete type holding information on species bounds.

CellFreeModelGenerationKit.parse_vff_model_document — Function

parse_vff_model_document(model::VLAbstractModelObject)::VLResult

Parse contents of the model file, extract information from various sections of the file and construct a rudimentary data dictionary.

Input arguments: model::VLAbstractModelObject - abstract Julia data object holding information for generating model code.

Output arguments: VLResult::VLResult - concrete type holding an intermediate data dictionary construction.