CellFreeModelGenerationKit.jl Documentation

CellFreeModelGenerationKit.jl Documentation

Index

Index

CellFreeModelGenerationKit.build_control_program_component

CellFreeModelGenerationKit.build_data_dictionary_program_component

CellFreeModelGenerationKit.build_julia_model_object

CellFreeModelGenerationKit.build_kinetics_program_component

CellFreeModelGenerationKit.generate

CellFreeModelGenerationKit.generate_default_project_file

CellFreeModelGenerationKit.generate_stoichiometric_matrix

CellFreeModelGenerationKit.parse_vff_bio_types_section

CellFreeModelGenerationKit.parse_vff_grn_section

CellFreeModelGenerationKit.parse_vff_metabolic_section

CellFreeModelGenerationKit.parse_vff_metabolic_section

CellFreeModelGenerationKit.parse_vff_model_document

CellFreeModelGenerationKit.parse_vff_model_document

CellFreeModelGenerationKit.parse_vff_sequence_section

CellFreeModelGenerationKit.parse_vff_sequence_section

CellFreeModelGenerationKit.parse_vff_species_bounds_section

CellFreeModelGenerationKit.parse_vff_species_bounds_section

CellFreeModelGenerationKit.read_model_document

CellFreeModelGenerationKit.reorder_molecular_symbol_array

CellFreeModelGenerationKit.reorder_reaction_symbol_array

CellFreeModelGenerationKit.build_control_program_component — Method

build_control_program_component(intermediate_dictionary::Dict{String,Any})::VLResult

CellFreeModelGenerationKit.build_data_dictionary_program_component — Method

build_data_dictionary_program_component(intermediate_dictionary::Dict{String,Any})::VLResult

CellFreeModelGenerationKit.build_julia_model_object — Method

build_julia_model_object(path_to_model_file::String, path_to_output_dir::String; 
    defaults_file_name::String="Defaults.toml", model_type::Symbol=:static)::VLResult

CellFreeModelGenerationKit.build_kinetics_program_component — Method

build_kinetics_program_component(intermediate_dictionary::Dict{String,Any})::VLResult

CellFreeModelGenerationKit.generate — Method

generate(julia_model_object::VLJuliaModelObject; 
    intermediate_representation_dictionary::Union{Nothing,Dict{String,Any}} = nothing, 
    logger::Union{Nothing,SimpleLogger} = nothing)

CellFreeModelGenerationKit.generate_default_project_file — Method

generate_default_project(path_to_project_dir::String)

Generates a default project structure which contains an empty model file and Defaults.toml file.

Inputs:
path_to_project_dir = path to where you want model code to be generated

CellFreeModelGenerationKit.generate_stoichiometric_matrix — Method

generate_stoichiometric_matrix(intermediate_dictionary::Dict{String,Any})::VLResult

CellFreeModelGenerationKit.parse_vff_bio_types_section — Method

parse_vff_bio_types_section(buffer::Array{String,1})::VLResult

CellFreeModelGenerationKit.parse_vff_grn_section — Method

parse_grn_section(buffer::Array{String,1})::VLResult

CellFreeModelGenerationKit.parse_vff_metabolic_section — Method

parse_vff_metabolic_section(buffer::Array{String,1})::VLResult

CellFreeModelGenerationKit.parse_vff_model_document — Method

parse_vff_model_document(model::VLAbstractModelObject)::VLResult

CellFreeModelGenerationKit.parse_vff_sequence_section — Method

parse_vff_sequence_section(buffer::Array{String,1})::VLResult

CellFreeModelGenerationKit.parse_vff_species_bounds_section — Method

parse_vff_species_bounds_section(buffer::Array{String,1}, metabolic_results_tuple::NamedTuple)::VLResult

CellFreeModelGenerationKit.read_model_document — Method

read_model_document(path_to_file::String; 
    strip_comments::Bool = true)::Array{String,1}

CellFreeModelGenerationKit.reorder_molecular_symbol_array — Method

reorder_molecular_symbol_array(symbol_array::Array{String,1}; 
    callback::Union{Function,Nothing} = nothing)::Array{String,1}

CellFreeModelGenerationKit.reorder_reaction_symbol_array — Method

reorder_reaction_symbol_array(symbol_array::Array{String,1}; 
    callback::Union{Function,Nothing} = nothing)::Array{String,1}