Advanced user controls

To extend the functionality of the existing code generator, advanced users can customize the following methods depending on their requirements.

Methods to generate model object

CellFreeModelGenerationKit.read_model_document — Function

read_model_document(path_to_file::String;
        strip_comments::Bool = true)::Array{String,1}

Read each line of the Network.vff input file and extract contents of the file. Comments in the file (beginning by // symbol) are excluded by default, which can be overridden by the user.

Input arguments: path_to_file::String - user-specified path where the Network.vff input file can be found and parsed strip_comments::Bool - if true, comment lines (beginning by ‘//’ symbol) are excluded by the parser (optional).

Output arguments: buffer::Array{String,1} - single array holding information extracted from the Network.vff file.

Missing docstring.

Missing docstring for generate_default_project. Check Documenter's build log for details.

Methods to build the data dictionary

CellFreeModelGenerationKit.generate_stoichiometric_matrix — Function

generate_stoichiometric_matrix(intermediate_dictionary::Dict{String,Any})::VLResult

Generate a stoichiometric matrix based on the biochemical model reaction network.

Input arguments: intermediate_dictionary::Dict{String,Any} - data dictionary containing the master reaction table and molecular species participating in the reactions.

Output arguments: VLResult::VLResult - concrete data type holding the generated stoichiometric matrix.

CellFreeModelGenerationKit.build_data_dictionary_program_component — Function

build_data_dictionary_program_component(intermediate_dictionary::Dict{String,Any})::VLResult

Generate Data.jl. Data.jl includes a method to build a default data dictionary. This dictionary will contain information about the model system along with default bounds useful for flux balance analysis.

Input arguments: intermediate_dictionary::Dict{String,Any} - data dictionary containing the master reaction table and molecular species participating in the reactions.

Output arguments: VLResult::VLResult - concrete data type holding the data dictionary.

CellFreeModelGenerationKit.build_control_program_component — Function

Function: build_control_program_component(intermediate_dictionary::Dict{String,Any})::VLResult

Generate Control.jl. Control.jl includes methods to calculate transcription and translation control functions at time t.

Input arguments: intermediate_dictionary::Dict{String,Any} - data dictionary containing the master reaction table and molecular species participating in the reactions.

Output arguments: VLResult::VLResult - concrete data type holding control values for each rate.

CellFreeModelGenerationKit.build_kinetics_program_component — Function

build_kinetics_dictionary_program_component(intermediate_dictionary::Dict{String,Any})::VLResult

Generate Kinetics.jl. Kinetics.jl includes methods to calculate transcription and translation rates based on biophysical model parameters.

Input arguments: intermediate_dictionary::Dict{String,Any} - data dictionary containing the master reaction table and molecular species participating in the reactions.

Output arguments: VLResult::VLResult - concrete data type holding all rates.

Methods to customize parsing strategy

CellFreeModelGenerationKit.parse_vff_model_document — Function

parse_vff_model_document(model::VLAbstractModelObject)::VLResult

Parse contents of the model file, extract information from various sections of the file and construct a rudimentary data dictionary.

Input arguments: model::VLAbstractModelObject - abstract Julia data object holding information for generating model code.

Output arguments: VLResult::VLResult - concrete type holding an intermediate data dictionary construction.

CellFreeModelGenerationKit.parse_vff_metabolic_section — Function

parse_vff_metabolic_section(buffer::Array{String,1})::VLResult

Parse contents of the metabolic section in the model file and extract all biochemical reactions and molecular species involved in the metabolic network. The metabolic section is written between the tags #METABOLISM::START and #METABOLISM::STOP.

Input arguments: buffer::Array{String,1} - single array holding information extracted from the model file.

Output arguments: VLResult::VLResult - concrete type holding information on reactions and species involved in metabolism.

CellFreeModelGenerationKit.parse_vff_sequence_section — Function

parse_vff_sequence_section(buffer::Array{String,1})::VLResult

Parse contents of the sequence section in the model file and extract amino acid and nucleotide sequences for protein and DNA, respectively. The sequence section is written between the tags #TXTL-SEQUENCE::START and #TXTL-SEQUENCE::STOP.

Input arguments: buffer::Array{String,1} - 1-D array holding information extracted from the model file.

Output arguments: VLResult::VLResult - concrete type holding protein and DNA sequence information.

CellFreeModelGenerationKit.parse_vff_bio_types_section — Function

parse_vff_bio_types_section(buffer::Array{String,1})::VLResult

Parse contents of the model file and tokenize biological types. The bio-types section is written between the tags #BIO-TYPE-PREFIXES::START and #BIO-TYPE-PREFIXES::STOP.

Input arguments: buffer::Array{String,1} - single array holding information extracted from the model file.

Output arguments: VLResult::VLResult - concrete type holding information on biological types.

CellFreeModelGenerationKit.parse_vff_species_bounds_section — Function

parse_vff_species_bounds_section(buffer::Array{String,1}, metabolic_results_tuple::NamedTuple)::VLResult

Parse contents of the model file and extract information on species bounds. The species bounds section is written between the tags #SPECIESBOUNDS::START and #SPECIESBOUNDS::STOP.

Input arguments: buffer::Array{String,1} - 1-D array holding information extracted from the model file.

Output arguments: VLResult::VLResult - concrete type holding information on species bounds.

CellFreeModelGenerationKit.parse_vff_grn_section — Function

parse_vff_grn_section(buffer::Array{String,1})::VLResult

Parse contents of the GRN section in the model file and extract information contained on gene regulation. The GRN section is written between the tags #GRN::START and #GRN::STOP.

Input arguments: buffer::Array{String,1} - 1-D array holding information extracted from the model file.

Output arguments: VLResult::VLResult - concrete type holding information on the gene regulatory network.

Callbacks

CellFreeModelGenerationKit.reorder_molecular_symbol_array — Function

reorder_molecular_symbol_array(symbol_array::Array{String,1};
      callback::Union{Function,Nothing} = nothing)::Array{String,1}

Reorder species symbol names alphabetically (default implementation) unless a callback function is passed, in which case sorting is done based on the user-defined routine.

Input arguments: symbol_array::Array{String,1} - array holding symbol names for all molecular species in the network. callback::Union{Function,Nothing} - user-defined function to reorder species symbol names (optional).

Output arguments: symbol_array::Array{String,1} - sorted molecular species symbol array.

CellFreeModelGenerationKit.reorder_reaction_symbol_array — Function

reorder_reaction_symbol_array(symbol_array::Array{String,1};
      callback::Union{Function,Nothing} = nothing)::Array{String,1}

Reorder reaction symbol names alphabetically (default implementation) unless a callback function is passed, in which case sorting is done based on the user-defined routine.

Input arguments: symbol_array::Array{String,1} - array holding symbol names for all reactions in the network. callback::Union{Function,Nothing} - user-defined function to reorder reaction names (optional).

Output arguments: symbol_array::Array{String,1} - sorted reaction symbol array.